@@ -1,7 +1,6 @@
 # This file is dual licensed under the terms of the Apache License, Version
 # 2.0, and the BSD License. See the LICENSE file in the root of this repository
 # for complete details.
-from __future__ import absolute_import, division, print_function
 
 import collections
 import itertools
@@ -9,19 +8,16 @@ import os
 import platform
 import sys
 
-import pretend
 import pytest
 
+from packaging._parser import Node
 from packaging.markers import (
-    Node,
     InvalidMarker,
-    UndefinedComparison,
-    UndefinedEnvironmentName,
     Marker,
+    UndefinedComparison,
     default_environment,
     format_full_version,
 )
-
 
 VARIABLES = [
     "extra",
@@ -65,13 +61,13 @@ class TestNode:
     def test_accepts_value(self, value):
         assert Node(value).value == value
 
-    @pytest.mark.parametrize("value", ["one", "two", None, 3, 5, []])
+    @pytest.mark.parametrize("value", ["one", "two"])
     def test_str(self, value):
         assert str(Node(value)) == str(value)
 
-    @pytest.mark.parametrize("value", ["one", "two", None, 3, 5, []])
+    @pytest.mark.parametrize("value", ["one", "two"])
     def test_repr(self, value):
-        assert repr(Node(value)) == "<Node({0!r})>".format(str(value))
+        assert repr(Node(value)) == f"<Node({str(value)!r})>"
 
     def test_base_class(self):
         with pytest.raises(NotImplementedError):
@@ -89,6 +85,11 @@ class TestOperatorEvaluation:
         with pytest.raises(UndefinedComparison):
             Marker("'2.7.0' ~= os_name").evaluate()
 
+    def test_allows_prerelease(self):
+        assert Marker('python_full_version > "3.6.2"').evaluate(
+            {"python_full_version": "3.11.0a5"}
+        )
+
 
 FakeVersionInfo = collections.namedtuple(
     "FakeVersionInfo", ["major", "minor", "micro", "releaselevel", "serial"]
@@ -96,51 +97,6 @@ FakeVersionInfo = collections.namedtuple(
 
 
 class TestDefaultEnvironment:
-    @pytest.mark.skipif(
-        hasattr(sys, "implementation"), reason="sys.implementation does exist"
-    )
-    def test_matches_expected_no_sys_implementation(self):
-        environment = default_environment()
-
-        assert environment == {
-            "implementation_name": "",
-            "implementation_version": "0",
-            "os_name": os.name,
-            "platform_machine": platform.machine(),
-            "platform_release": platform.release(),
-            "platform_system": platform.system(),
-            "platform_version": platform.version(),
-            "python_full_version": platform.python_version(),
-            "platform_python_implementation": platform.python_implementation(),
-            "python_version": ".".join(platform.python_version_tuple()[:2]),
-            "sys_platform": sys.platform,
-        }
-
-    @pytest.mark.skipif(
-        not hasattr(sys, "implementation"), reason="sys.implementation does not exist"
-    )
-    def test_matches_expected_deleted_sys_implementation(self, monkeypatch):
-        monkeypatch.delattr(sys, "implementation")
-
-        environment = default_environment()
-
-        assert environment == {
-            "implementation_name": "",
-            "implementation_version": "0",
-            "os_name": os.name,
-            "platform_machine": platform.machine(),
-            "platform_release": platform.release(),
-            "platform_system": platform.system(),
-            "platform_version": platform.version(),
-            "python_full_version": platform.python_version(),
-            "platform_python_implementation": platform.python_implementation(),
-            "python_version": ".".join(platform.python_version_tuple()[:2]),
-            "sys_platform": sys.platform,
-        }
-
-    @pytest.mark.skipif(
-        not hasattr(sys, "implementation"), reason="sys.implementation does not exist"
-    )
     def test_matches_expected(self):
         environment = default_environment()
 
@@ -155,32 +111,6 @@ class TestDefaultEnvironment:
         assert environment == {
             "implementation_name": sys.implementation.name,
             "implementation_version": iver,
-            "os_name": os.name,
-            "platform_machine": platform.machine(),
-            "platform_release": platform.release(),
-            "platform_system": platform.system(),
-            "platform_version": platform.version(),
-            "python_full_version": platform.python_version(),
-            "platform_python_implementation": platform.python_implementation(),
-            "python_version": ".".join(platform.python_version_tuple()[:2]),
-            "sys_platform": sys.platform,
-        }
-
-    @pytest.mark.skipif(
-        hasattr(sys, "implementation"), reason="sys.implementation does exist"
-    )
-    def test_monkeypatch_sys_implementation(self, monkeypatch):
-        monkeypatch.setattr(
-            sys,
-            "implementation",
-            pretend.stub(version=FakeVersionInfo(3, 4, 2, "final", 0), name="linux"),
-            raising=False,
-        )
-
-        environment = default_environment()
-        assert environment == {
-            "implementation_name": "linux",
-            "implementation_version": "3.4.2",
             "os_name": os.name,
             "platform_machine": platform.machine(),
             "platform_release": platform.release(),
@@ -217,7 +147,7 @@ class TestMarker:
     @pytest.mark.parametrize(
         "marker_string",
         [
-            "{0} {1} {2!r}".format(*i)
+            "{} {} {!r}".format(*i)
             for i in itertools.product(VARIABLES, OPERATORS, VALUES)
         ]
         + [
@@ -235,6 +165,7 @@ class TestMarker:
             "python_version",
             "(python_version)",
             "python_version >= 1.0 and (python_version)",
+            '(python_version == "2.7" and os_name == "linux"',
         ],
     )
     def test_parses_invalid(self, marker_string):
@@ -276,21 +207,59 @@ class TestMarker:
             ),
         ],
     )
-    def test_str_and_repr(self, marker_string, expected):
+    def test_str_repr_eq_hash(self, marker_string, expected):
         m = Marker(marker_string)
         assert str(m) == expected
-        assert repr(m) == "<Marker({0!r})>".format(str(m))
+        assert repr(m) == f"<Marker({str(m)!r})>"
+        # Objects created from the same string should be equal.
+        assert m == Marker(marker_string)
+        # Objects created from the equivalent strings should also be equal.
+        assert m == Marker(expected)
+        # Objects created from the same string should have the same hash.
+        assert hash(Marker(marker_string)) == hash(Marker(marker_string))
+        # Objects created from equivalent strings should also have the same hash.
+        assert hash(Marker(marker_string)) == hash(Marker(expected))
 
-    def test_extra_with_no_extra_in_environment(self):
-        # We can't evaluate an extra if no extra is passed into the environment
-        m = Marker("extra == 'security'")
-        with pytest.raises(UndefinedEnvironmentName):
-            m.evaluate()
+    @pytest.mark.parametrize(
+        ("example1", "example2"),
+        [
+            # Test trivial comparisons.
+            ('python_version == "2.7"', 'python_version == "3.7"'),
+            (
+                'python_version == "2.7"',
+                'python_version == "2.7" and os_name == "linux"',
+            ),
+            (
+                'python_version == "2.7"',
+                '(python_version == "2.7" and os_name == "linux")',
+            ),
+            # Test different precedence.
+            (
+                'python_version == "2.7" and (os_name == "linux" or '
+                'sys_platform == "win32")',
+                'python_version == "2.7" and os_name == "linux" or '
+                'sys_platform == "win32"',
+            ),
+        ],
+    )
+    def test_different_markers_different_hashes(self, example1, example2):
+        marker1, marker2 = Marker(example1), Marker(example2)
+        # Markers created from strings that are not equivalent should differ.
+        assert marker1 != marker2
+        # Different Marker objects should have different hashes.
+        assert hash(marker1) != hash(marker2)
+
+    def test_compare_markers_to_other_objects(self):
+        # Markers should not be comparable to other kinds of objects.
+        assert Marker("os_name == 'nt'") != "os_name == 'nt'"
+
+    def test_environment_assumes_empty_extra(self):
+        assert Marker('extra == "im_valid"').evaluate() is False
 
     @pytest.mark.parametrize(
         ("marker_string", "environment", "expected"),
         [
-            ("os_name == '{0}'".format(os.name), None, True),
+            (f"os_name == '{os.name}'", None, True),
             ("os_name == 'foo'", {"os_name": "foo"}, True),
             ("os_name == 'foo'", {"os_name": "bar"}, False),
             ("'2.7' in python_version", {"python_version": "2.7.5"}, True),
@@ -320,6 +289,14 @@ class TestMarker:
             ),
             ("extra == 'security'", {"extra": "quux"}, False),
             ("extra == 'security'", {"extra": "security"}, True),
+            ("extra == 'SECURITY'", {"extra": "security"}, True),
+            ("extra == 'security'", {"extra": "SECURITY"}, True),
+            ("extra == 'pep-685-norm'", {"extra": "PEP_685...norm"}, True),
+            (
+                "extra == 'Different.punctuation..is...equal'",
+                {"extra": "different__punctuation_is_EQUAL"},
+                True,
+            ),
         ],
     )
     def test_evaluates(self, marker_string, environment, expected):
@@ -329,7 +306,7 @@ class TestMarker:
     @pytest.mark.parametrize(
         "marker_string",
         [
-            "{0} {1} {2!r}".format(*i)
+            "{} {} {!r}".format(*i)
             for i in itertools.product(PEP_345_VARIABLES, OPERATORS, VALUES)
         ]
         + [
@@ -343,7 +320,7 @@ class TestMarker:
     @pytest.mark.parametrize(
         ("marker_string", "environment", "expected"),
         [
-            ("os.name == '{0}'".format(os.name), None, True),
+            (f"os.name == '{os.name}'", None, True),
             ("sys.platform == 'win32'", {"sys_platform": "linux2"}, False),
             ("platform.version in 'Ubuntu'", {"platform_version": "#39"}, False),
             ("platform.machine=='x86_64'", {"platform_machine": "x86_64"}, True),
@@ -367,7 +344,7 @@ class TestMarker:
     @pytest.mark.parametrize(
         "marker_string",
         [
-            "{0} {1} {2!r}".format(*i)
+            "{} {} {!r}".format(*i)
             for i in itertools.product(SETUPTOOLS_VARIABLES, OPERATORS, VALUES)
         ]
         + [
@@ -382,3 +359,12 @@ class TestMarker:
         marker_string = "python_implementation=='Jython'"
         args = [{"platform_python_implementation": "CPython"}]
         assert Marker(marker_string).evaluate(*args) is False
+
+    def test_extra_str_normalization(self):
+        raw_name = "S_P__A_M"
+        normalized_name = "s-p-a-m"
+        lhs = f"{raw_name!r} == extra"
+        rhs = f"extra == {raw_name!r}"
+
+        assert str(Marker(lhs)) == f'"{normalized_name}" == extra'
+        assert str(Marker(rhs)) == f'extra == "{normalized_name}"'
