@@ -1,18 +1,18 @@
-# -*- coding: utf-8 -*-
 # This file is dual licensed under the terms of the Apache License, Version
 # 2.0, and the BSD License. See the LICENSE file in the root of this repository
 # for complete details.
-from __future__ import absolute_import, division, print_function
 
 import os
-import sys
 
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath("."))
+# -- Project information loading ----------------------------------------------
+
+ABOUT = {}
+_BASE_DIR = os.path.join(os.path.dirname(__file__), os.pardir)
+with open(os.path.join(_BASE_DIR, "packaging", "__about__.py")) as f:
+    exec(f.read(), ABOUT)
 
 # -- General configuration ----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions  coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
@@ -21,93 +21,48 @@ extensions = [
     "sphinx.ext.doctest",
     "sphinx.ext.extlinks",
     "sphinx.ext.intersphinx",
-    "sphinx.ext.viewcode",
 ]
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
-
-# The suffix of source filenames.
-source_suffix = ".rst"
-
-# The master toctree document.
-master_doc = "index"
 
 # General information about the project.
 project = "Packaging"
+version = ABOUT["__version__"]
+release = ABOUT["__version__"]
+copyright = ABOUT["__copyright__"]
 
-# The version info for the project you're documenting, acts as replacement for
-# |version| and |release|, also used in various other places throughout the
-# built documents.
-#
-
-base_dir = os.path.join(os.path.dirname(__file__), os.pardir)
-about = {}
-with open(os.path.join(base_dir, "packaging", "__about__.py")) as f:
-    exec(f.read(), about)
-
-version = release = about["__version__"]
-copyright = about["__copyright__"]
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-exclude_patterns = ["_build"]
-
-extlinks = {
-    "issue": ("https://github.com/pypa/packaging/issues/%s", "#"),
-    "pull": ("https://github.com/pypa/packaging/pull/%s", "PR #"),
-}
 # -- Options for HTML output --------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 
 html_theme = "furo"
-html_title = "packaging"
+html_title = project
 
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["_static"]
+# -- Options for autodoc ----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#configuration
 
-# Output file base name for HTML help builder.
-htmlhelp_basename = "packagingdoc"
+autodoc_member_order = "bysource"
+autodoc_preserve_defaults = True
 
+# Automatically extract typehints when specified and place them in
+# descriptions of the relevant function/method.
+autodoc_typehints = "description"
 
-# -- Options for LaTeX output -------------------------------------------------
+# Don't show class signature with the class' name.
+autodoc_class_signature = "separated"
 
-latex_elements = {}
+# -- Options for extlinks -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/extensions/extlinks.html#configuration
 
-# Grouping the document tree into LaTeX files. List of tuples
-# (source start file, target name, title, author, documentclass [howto/manual])
-latex_documents = [
-    ("index", "packaging.tex", "Packaging Documentation", "Donald Stufft", "manual")
-]
+extlinks = {
+    "issue": ("https://github.com/pypa/packaging/issues/%s", "#%s"),
+    "pull": ("https://github.com/pypa/packaging/pull/%s", "PR #%s"),
+}
 
-# -- Options for manual page output -------------------------------------------
+# -- Options for intersphinx ----------------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html#configuration
 
-# One entry per manual page. List of tuples
-# (source start file, name, description, authors, manual section).
-man_pages = [("index", "packaging", "Packaging Documentation", ["Donald Stufft"], 1)]
-
-# -- Options for Texinfo output -----------------------------------------------
-
-# Grouping the document tree into Texinfo files. List of tuples
-# (source start file, target name, title, author,
-#  dir menu entry, description, category)
-texinfo_documents = [
-    (
-        "index",
-        "packaging",
-        "Packaging Documentation",
-        "Donald Stufft",
-        "packaging",
-        "Core utilities for Python packages",
-        "Miscellaneous",
-    )
-]
-
-# Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {"https://docs.python.org/": None}
-
-epub_theme = "epub"
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3/", None),
+    "pypug": ("https://packaging.python.org/", None),
+}
